:package: Chore: move ThemeWrapper from pages to components folder, remove react logo
